+++ src/main/java/io/xlorey/FluxLoader/client/core/WidgetManager.java
@@ -1,10 +1,11 @@
 package io.xlorey.FluxLoader.client.core;
 
-import imgui.ImGui;
-import imgui.ImGuiIO;
+import imgui.*;
 import imgui.gl3.ImGuiImplGl3;
 import imgui.glfw.ImGuiImplGlfw;
 import io.xlorey.FluxLoader.client.api.TextTools;
+import io.xlorey.FluxLoader.client.fonts.FontData;
+import io.xlorey.FluxLoader.client.fonts.FontAwesomeIcons;
 import io.xlorey.FluxLoader.client.ui.ScreenWidget;
 import io.xlorey.FluxLoader.client.ui.pluginMenu.MainMenuButton;
 import io.xlorey.FluxLoader.client.ui.pluginMenu.PluginMenu;
@@ -17,7 +18,14 @@ import zombie.core.opengl.RenderThread;
 import zombie.gameStates.MainScreenState;
 import zombie.ui.UIFont;
 
+import java.io.IOException;
+import java.io.InputStream;
+import java.net.URL;
+import java.nio.file.Files;
+import java.nio.file.Paths;
 import java.util.ArrayList;
+import java.util.HashMap;
+import java.util.Map;
 
 /**
  * Custom UI management system
@@ -54,6 +62,16 @@ public class WidgetManager {
     private static boolean isMouseCapture = false;
 
     /**
+     * Custom font data fonts
+     */
+    private static final Map<String, FontData> customFontsData = new HashMap<>();
+
+    /**
+     * Loaded custom fonts
+     */
+    private static final Map<String, ImFont> customFonts = new HashMap<>();
+
+    /**
      * Change mouse capture state for ImGui widgets. Used only inside widgets via the captureMouseFocus method
      */
     public static void captureMouse() {
@@ -84,6 +102,107 @@ public class WidgetManager {
     }
 
     /**
+     * Initializes and adds fonts to ImGui.
+     * Adds a standard font, as well as fonts for Cyrillic, Japanese characters and FontAwesome icons.
+     * Fonts are merged to provide support for different glyph ranges in a single font.
+     * @param io An ImGuiIO object used to configure fonts.
+     */
+    private static void initFonts(final ImGuiIO io) {
+        // Add default font for latin glyphs
+        io.getFonts().addFontDefault();
+
+        // Create a builder for glyph ranges
+        final ImFontGlyphRangesBuilder rangesBuilder = new ImFontGlyphRangesBuilder();
+        rangesBuilder.addRanges(io.getFonts().getGlyphRangesDefault()); // Latin
+        rangesBuilder.addRanges(io.getFonts().getGlyphRangesCyrillic()); // Cyrillic
+        rangesBuilder.addRanges(io.getFonts().getGlyphRangesJapanese()); // Japanese
+        rangesBuilder.addRanges(FontAwesomeIcons._IconRange); // FontAwesome icons
+
+        // Create a configuration for additional fonts
+        final ImFontConfig fontConfig = new ImFontConfig();
+        fontConfig.setMergeMode(true); // Enable merge mode
+
+        // Build glyph ranges
+        final short[] glyphRanges = rangesBuilder.buildRanges();
+
+        // Add custom fonts with merged glyph ranges
+        io.getFonts().addFontFromMemoryTTF(loadFont("Tahoma.ttf"), 14, fontConfig, glyphRanges); // Cyrillic glyphs
+        io.getFonts().addFontFromMemoryTTF(loadFont("NotoSansCJKjp-Medium.otf"), 14, fontConfig, glyphRanges); // Japanese glyphs
+        io.getFonts().addFontFromMemoryTTF(loadFont("fa-regular-400.ttf"), 14, fontConfig, glyphRanges); // FontAwesome regular
+        io.getFonts().addFontFromMemoryTTF(loadFont("fa-solid-900.ttf"), 14, fontConfig, glyphRanges); // FontAwesome solid
+
+        // Add custom fonts
+        for (Map.Entry<String, FontData> entry : customFontsData.entrySet()) {
+            String fontName = entry.getKey();
+            FontData fontData = entry.getValue();
+
+            // Use the same fontConfig and glyphRanges for custom fonts
+            ImFont font = io.getFonts().addFontFromMemoryTTF(fontData.getFontData(), fontData.getSize(), fontConfig, glyphRanges);
+
+            customFonts.put(fontName, font);
+        }
+
+        io.getFonts().build();
+        fontConfig.destroy();
+    }
+
+    /**
+     * Gets a custom font by its name.
+     * @param fontName The name of the font to get.
+     * @return An ImFont object representing the custom font, or null,
+     * if a font with the specified name is not found.
+     */
+    public static ImFont getCustomFont(String fontName) {
+        return customFonts.getOrDefault(fontName, null);
+    }
+
+    /**
+     * Loads a font from a file.
+     * @param name The name of the font file.
+     * @return A byte array containing the font data.
+     */
+    private static byte[] loadFont(String name) {
+        try {
+            String fontPath = "io/xlorey/FluxLoader/media/fonts/" + name;
+            return Files.readAllBytes(Paths.get(fontPath));
+        } catch (IOException e) {
+            throw new RuntimeException("Error loading font: " + name, e);
+        }
+    }
+
+    /**
+     * Loads font data from the specified URL.
+     * @param url URL to the font file.
+     * @return A byte array containing the font data.
+     */
+    private static byte[] loadFont(URL url) {
+        try (InputStream is = url.openStream()) {
+            return is.readAllBytes();
+        } catch (IOException e) {
+            throw new RuntimeException("Error loading font from URL: " + url, e);
+        }
+    }
+
+    /**
+     * Loads and adds a font to ImGui from the resource specified by path.
+     * This method is intended for use with fonts embedded in the application JAR file.
+     * @param path Path to the font resource in the JAR file.
+     * @param fontName Font name (Its ID)
+     * @param size Font size.
+     */
+    public static void addCustomFont(URL path, String fontName, float size) {
+        final ImFontConfig fontConfig = new ImFontConfig();
+
+        byte[] fontData = loadFont(path);
+
+        FontData font = new FontData(fontName, size, fontData);
+
+        customFontsData.put(fontName, font);
+
+        fontConfig.destroy();
+    }
+
+    /**
      * Initializing the ImGui context
      */
     private static void initImGui() {
@@ -92,6 +211,8 @@ public class WidgetManager {
         ImGuiIO io = ImGui.getIO();
         // Preventing UI Elements from Saving State
         io.setIniFilename(null);
+
+        initFonts(io);
 
         imGuiGlfw.init(windowHandler, true);
         imGuiGl3.init("#version 130");