+++ changelog.md
@@ -1,5 +1,6 @@
 # FluxLoader v0.8.0
 - Added missing `Configuration` data types
+- Added new fonts for `ImGui` and the ability to load custom ones
 
 # FluxLoader v0.7.1
 - Fixed creation of plugins folder